264. API Authorization #208

+++ app/Http/Controllers/Api/V1/PostCommentController.php
@@ -78,6 +78,7 @@ class PostCommentController extends Controller
      */
     public function update(BlogPost $post, Comment $comment, StoreComment $request)
     {
+        $this->authorize($comment);
         $comment->content = $request->input('content');
         $comment->save();
 
@@ -92,6 +93,7 @@ class PostCommentController extends Controller
      */
     public function destroy(BlogPost $post, Comment $comment)
     {
+        $this->authorize($comment);
         $comment->delete();
 
         return response()->noContent();
